// src/apis/models/FireBytes_Unified_Wave_InterfaceServices_Dtos_InterfaceServiceEditDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_InterfaceServices_Dtos_InterfaceServiceEditDto = {
    id?: string | null;
    serviceName?: string | null;
    serviceIp?: string | null;
    servicePort?: number;
    serviceId?: string | null;
    servicePassword?: string | null;
    description?: string | null;
    interfaceBrandId?: string | null;
    isEnableSSL?: boolean | null;
};