// src/apis/models/FireBytes_Unified_Wave_InterfaceServices_Dtos_IpPortRepeatVerifyInput.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type FireBytes_Unified_Wave_InterfaceServices_Dtos_IpPortRepeatVerifyInput = {
    id?: string | null;
    serviceIp?: string | null;
    servicePort?: number;
};